+++ plugins/flux/src/components/page/PageToolbarItemClusters.tsx
@@ -25,6 +25,11 @@ const ToolbarItemClusters: React.FunctionComponent<IToolbarItemClustersProps> = 
       onClear={(): void => selectCluster('')}
       selections={selectedCluster}
       isOpen={showOptions}
+      onFilter={(e: React.ChangeEvent<HTMLInputElement> | null, value: string): React.ReactElement[] =>
+        clusters
+          .filter((c) => !value || c.includes(value))
+          .map((cluster: string) => <SelectOption key={cluster} value={cluster} />)
+      }
     >
       {clusters.map((cluster, index) => (
         <SelectOption key={index} value={cluster} />