+++ plugins/resources/src/components/page/PageToolbarItemResources.tsx
@@ -27,6 +27,19 @@ const ToolbarItemResources: React.FunctionComponent<IToolbarItemResourcesProps> 
       onClear={(): void => selectResource('')}
       selections={selectedResources}
       isOpen={showOptions}
+      onFilter={(e: React.ChangeEvent<HTMLInputElement> | null, value: string): React.ReactElement[] =>
+        Object.keys(resources)
+          .filter((key) => !value || key.includes(value))
+          .map((key) => (
+            <SelectOption
+              key={key}
+              value={key}
+              description={resources[key].isCRD ? `${resources[key].resource}.${resources[key].path}` : undefined}
+            >
+              {resources[key].title}
+            </SelectOption>
+          ))
+      }
     >
       {Object.keys(resources).map((key) => (
         <SelectOption